big refactoring to move the send/receive functions into App module
