+++ hwormhole.cabal
@@ -44,6 +44,8 @@ library
                      , memory          >= 0.14.15 && < 1.0
                      , network         >= 2.7     && < 3
                      , protolude       >= 0.2.1   && < 1.0
+                     , haskeline
+                     , random
                      , hex             >= 0.1.2   && < 1.0
                      , saltine         == 0.1.0.1 && < 1.0
                      , cryptonite      >= 0.24    && < 1.0
@@ -65,13 +67,9 @@ executable hwormhole-exe
                      , bytestring
                      , aeson           >=1.4      && <2
                      , binary
-                     , haskeline
                      , hwormhole
-                     , magic-wormhole
                      , optparse-applicative
                      , protolude
-                     , random
-                     , spake2          >= 0.4
                      , text
   hs-source-dirs:      cmd
   default-language:    Haskell2010
